CA-156 Fix document.title regression
This commit fixes an issue where the BCV would not show up in the document title when the app launches on a route other than the alignment editor. This issue likely wasn't noticed earlier because the app used to open by default on the alignment editor, and now it opens by default on the projects route.