CA-166 Fix QA Defect Related to Deleting a Project
This commit fixes this issue:

On client A, Create and upload a project to the server.
On client B, refresh remote projects to pull in the new project BUT don’t download it
On client B, use the 3 button menu to delete the server project - instead of being removed from the UI, it becomes a local only project and allows user to upload it to the server.