CA-207 Refactor creation of database
This commit fixes the following scenario: A user creates a new project and uploads a target .tsv file that does not contain a column header of 'required'. This commit makes it so the default behavior in this case is that the required field gets populated with a 1 in the database.